+++ solutions/ColumnFormatter/src/webparts/columnFormatter/loc/mystrings.d.ts
@@ -8,6 +8,7 @@ declare interface IColumnFormatterWebPartStrings {
   PropertyVisibleOn: string;
   PropertyVisibleOff: string;
   PropertyMiniMapLabel: string;
+  PropertyIndentGuidesLabel: string;
 
   //General
   FeatureUnavailableFromLocalWorkbench: string;